Add shadowban command (#754)
* Rewrite lastrun file logic

Changed lastrun file from a custom format to a JSON file. If a `lastrun.json` file doesn't exist yet, the old format will be automatically migrated.

* Add shadowban command

The command does nothing but add the shadowban to the lastrun file.

* Add shadowban module

This module actually does all the work for the shadowban, removing all comments, attachments, and bug reports from shadowbanned users.

I've modified the `ExecutionTimeframe` so that it also stores all shadowbans; as a consequence, the module needs to be able to access the entire timeframe instead of just the `lastRunTime`. I've chosen to add an alternative way to implement modules if more details are needed; it might make sense to migrate to this for all modules in the future; or alternatively do a refactor so that a module can access anything non-static it needs from its invocation arguments.

* Add debug option to not ignore commands sent by bot user